aula1/src/requirements.txt
openai==1.81.0
python-dotenv==1.1.0
streamlit==1.45.1
google-genai==1.16.1